requirements.txt
# django
Django==3.2.18
django-cors-headers==4.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-environ==0.11.2

# postgresql
psycopg2-binary==2.9.9

# celery
celery==5.3.4
django-celery-results==2.5.1

# redis
redis==5.0.1
